v1.1.0 (wip)
v1.1.0 will add an option in `General` Settings to where it allows you to switch the url from “https://youtu.be/videoID?t=timestamp” to “https://www.youtube.com/watch?v=videoID&t=timestamp” since it would be a generous way of changing the url format. And I probably might include another option to where it filters out the `?t=timestamp` parameter if some plan on using it as a “Share” Button.